refactor: Migrate asset system to using templates
This change should make it more trivial to implement
the envy-designer. It allows us to operate off of the templates
instead of live-nodes, which means that we can access the fields
directly (and safely) without violating invariants for the runtime
system.

It also enables us to implement the serialization/deserialization
way simpler. For example, just compare the v010 deserialization
and serialization to the new implementation. Much simpler.

Finally, this allows us to implement a way to sync state from a source
of truth (the template) to the live layout tree. This enables
animation previewing and resetting.